Add wrapper for whole USART to firmware-lib
